move neigborfiles func from worksapceUtils to crossfileContextUtil
